requirements.txt
anyio==4.9.0
blinker==1.9.0
click==8.2.1
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-wtf==1.2.2
greenlet==3.2.2
h11==0.16.0
httptools==0.6.4
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pyyaml==6.0.2
sniffio==1.3.1
sqlalchemy==2.0.41
typing-extensions==4.13.2
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
werkzeug==3.1.3
whitenoise==6.9.0
wtforms==3.2.1
